+++ server-data/resources/[esx_addons]/esx-radialmenu/fxmanifest.lua
@@ -13,16 +13,12 @@ shared_scripts {
 
 client_scripts {
     'locales/*.lua',
-    'client/main.lua',
-    'client/clothing.lua',
-    'client/trunk.lua',
-    'client/stretcher.lua'
+    'client/*.lua',
 }
 
 server_scripts {
     'locales/*.lua',
-    'server/trunk.lua',
-    'server/stretcher.lua'
+    'server/*.lua',
 }
 
 files {
@@ -31,4 +27,3 @@ files {
     'html/js/main.js',
     'html/js/RadialMenu.js',
 }
-