+++ server-data/resources/[esx_addons]/esx-radialmenu/server/trunk.lua
@@ -12,15 +12,15 @@ RegisterNetEvent('esx-trunk:server:KidnapTrunk', function(targetId, closestVehic
     TriggerClientEvent('esx-trunk:client:KidnapGetIn', targetId, closestVehicle)
 end)
 
-ESX.RegisterServerCallback('esx-trunk:server:getTrunkBusy', function(src, cb, plate)
+ESX.RegisterServerCallback('esx-trunk:server:getTrunkBusy', function(_, cb, plate)
     if trunkBusy[plate] then cb(true) return end
     cb(false)
 end)
 
-ESX.RegisterCommand({"getintrunk"}, 'user', function(xPlayer, args, showError)
+ESX.RegisterCommand({"getintrunk"}, 'user', function(xPlayer)
     TriggerClientEvent('esx-trunk:client:GetIn', xPlayer.source)
 end, false, {help =  _U("getintrunk_command_desc")})
 
-ESX.RegisterCommand({"putintrunk"}, 'user', function(xPlayer, args, showError)
+ESX.RegisterCommand({"putintrunk"}, 'user', function(xPlayer, args)
     TriggerClientEvent('esx-trunk:server:KidnapTrunk', xPlayer.source)
 end, false, {help =  _U("putintrunk_command_desc")})