+++ .github/CHANGELOG.md
@@ -51,3 +51,8 @@
 26. [esx_status/server]: fix: unused argument @bitpredator
 27. [esx_addoninventory]: chore: check for nil society @bitpredator
 28. [npwd]: chore: npwd update to version 1.8.6 @bitpredator
+29. [bpt_deliveries]: chore: clean up the code @bitpredator
+30. [esx-radialmenu]: chore:  fix lint error + clean up the code @bitpredator
+31. [esx-qalle-jail]: fix: fix lint error @bitpredator
+32. [esx_multicaracter]: fix: fix lint error @bitpredator
+33. [esx_dmvschool]: fix: fix lint error @bitpredator